# Avoid re-hashing (twice!) in VacantEntry::insert