Anecdotal local benchmarks:

```
# Before
├─ u64_insert_intmap_entry               67.45 µs      │ 164.8 µs      │ 67.79 µs      │ 69.32 µs      │ 100     │ 100
# After
├─ u64_insert_intmap_entry               44.83 µs      │ 155.7 µs      │ 45.22 µs      │ 47.24 µs      │ 100     │ 100
```