@jrose-signal 2024-12-13
Does not reproduce for me, but I'm on an Apple M1 Max (aarch64), and the calling convention *is* different between aarch64 and x86_64. Feel free to steal the PR from me if you get a chance to look at it sooner than I do.