@jakoschiko 2024-12-14
I assume that the additional mutable `bool` prevents the compiler from inlining `increase_cache_if_needed` in `insert`, even if it's annotated with `#[inline]`. Using `#[inline(always)]` instead of `#[inline]` fixes the regression on my machine. Can you test that this doesn't create any regression on your machine?