@jakoschiko 2024-12-14
Also it looks like `integration_tests/Cargo.lock` is not up-to-date. Can you commit the updated file?