@jrose-signal 2024-12-16
I thought of something else immediately, heh. This is not as good—neither as clean, nor as fast—but it avoids touching `ensure_load_rate` at all, and it's still faster than the state of things today. What do you think?