@jakoschiko 2024-12-16
Ah nice. It works.

But this is also fast on my machine:

```rust
#[inline]
fn increase_cache_if_needed(&mut self) -> bool {
    let initial_cache_len = self.cache.len();

    // Handle empty cache to prevent division by zero.
    if initial_cache_len == 0 {
        self.increase_cache()
    }

    // Tried using floats here but insert performance tanked.
    while ((self.count * 1000) / self.cache.len()) > self.load_factor {
        self.increase_cache();
    }

    return initial_cache_len != self.cache.len()
}
```

I assume that the additional assignment inside the loop was the problem. Can you try this?